ci: Create GitHub Release automatically
Add a GitHub workflow to create the GitHub Release automatically when
a version is tagged. Also use the GitHub feature to create the CHANGELOG
automatically. When creating the GitHub release the contents of the file
RELEASE_NOTES.md will be used as release notes, so it should only
contain an explanation on the release main changes and be user-facing.

Also add the release notes for the current release.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>