Create contributing guide
Add a CONTRIBUTING.md guide explaining how to build, test, setup
a development environment and release the library.

Also mention the contributing guide in the README.md and add a
RELEASE_NOTES template, so when releasing new versions is easy to clear
the release notes by just copying the template.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>