Store the original Component and Connection in the component graph
For some reason the Component and Connection objects were converted to
`dict` to store them in the component graph. This requires re-building
the objects when they need to be extracted, which is wasteful.

It also means that type information is lost, and in the future, when
adopting the microgrid API client using v0.17, we have a class hierarchy
for components instead of relaying in the category and type, so we would
even need to store the type if we wanted to rebuild the objects.

Having nested objects is also an issue, as then the nested properties
were kept as `dict`, adding an inconsistency and making it type-unsafe.

With this change we are able to get rid of a hack to retrieve the grid
metadata.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>