Update bound-streaming channels to resend latest to new receivers
The `SystemBoundsTracker`s send out a new value only when there's a
change in the system bounds.  This means when a new receiver is
created, it could wait for a long time before receiving the bounds.

This change allows new receivers to get the latest state when
requested.

This is already the case for battery bounds, and needed only for PV
and EV charger bounds.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>