Drop support for using `marshmallow_dataclas.dataclass` directly
`marshmallow_dataclass.dataclass` is intended to be used only when using
`my_dataclass.Schema` to get the schema. But using this is not very
convenient when using type hints as they are not well supported by
`marshmallow`, as the `load()` function can't have hints.

This is actually why `load_config()` exists in the first place, so we
are using `class_schema()` instead, so we don't really need that our
types are decorated with `marshmallow_dataclass`, we can use the
built-in `dataclass` instead, we just need to add the appropriate
metadata if we want more complex validation.

Using `class_shema()` is also necessary to be able to pass a
`base_schema`, which we'll need when we want to use schemas with
custom fields, like the ones provided by `frequenz.quantities`.

Because of this, we just drop support for
`marshmallow_dataclass.dataclass` and we'll require that built-in
dataclasses are used in the future.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>