Revert "Add an option to wait for the first configuration"
This feature is adding too much complexity, just to make it slightly
easier to debug a very specific case, when a receiver gets stuck without
being able to receive any messages at all.

We will improve logging instead, so we can still debug this case without
adding this complexity to the API.

This also allows us to make the `new_receiver` method sync.

This reverts commit 29725c4a87095ce82ffec3e294342e8115802299.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>