Remove support for receiving raw mapping as configuration
This seems to be a very niche feature that adds quite a bit of
complexity. Users than need this kind of raw access can just get
a receiver from the `config_channel` themselves and do the
processing they need.

Now the `schema` is required, was renamed to `config_class` for extra
clarity and is a positional-only argument.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>