Improve handling of unknown fields in the configuration
If `unkown` is not specified in the `marshmallow_load_kwargs`, it will
default to `marshmallow.EXCLUDE` instead of `marshmallow.RAISE`, as this
is what most users will need when loading to a dataclass.

But when `marshmallow.EXCLUDE` is used, a warning will be logged if
there are fields in the configuration that are being excluded, so the
user can still be aware of it and catch typos in the configuration file.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>