Document and fix how to implement custom stopping logic
The `BackgroundService` should not return early if there are no tasks
in `stop()`, as sub-classes could need to implement a custom stopping
logic. In this case it should be enough to override the `cancel()` and
`wait()` methods, and the `is_running` property.

We also add an example of how this could be done.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>