Validate and load configurations using a config dataclass
This commit adds support for validating configurations with a schema.
This is useful to ensure the configuration is correct and to receive it
as an instance of a dataclass.

The `new_receiver` method now requires a `config_class` argument that is
used to validate the configuration. If the configuration doesn't pass
the validation, it will be ignored and an error will be logged.

The schema is expected to be a dataclass, which is used to create a
marshmallow schema to validate the configuration. To customize the
schema, you can use `marshmallow_dataclass` to specify extra metadata.

Validation errors are logged and the `ValidationError` instance is sent
to the receiver.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>