Change debug logs to info
Updating loggers config is not that often, and to be able to debug with
some confidence, is good to see exactly which loggers were enabled with
which levels, so we convert logs about configuring the loggers from
debug to info.

Now that we are showing how loggers are being configured, there is no
need to show the whole new configuration, which could be pretty lengthy.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>