Add a new utility function to wait for the first config
This function has a default timeout, so waiting for the first
configuration can be automatically aborted when it takes too long.

The method will also reports some progress, logging errors if there were
issues receiving or loading the configuration, or if the configuration
is missing, setting a logging baseline for all actors.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>