Allow passing a single configuration file
Both the `ConfigManager` and the `ConfigManagingActor` can now take a
single configuration file, both as a `str` or a `Path`.

This also fixes a bug where `str` was still accepted, but taken as a
`Sequence[str]`. This is a known issue with Python typing:
https://github.com/python/mypy/issues/11001

And while we are it, we now also explicitly reject empty sequences,
raising a `ValueError` when one is passed.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>